add a max URL to download before saving into the record
in order to complete the wait function, we added a function to browse the list in the reverse way
but this operation, on the whole buffer took a long time. Thus, to be sure to save your work, we add
a max_url_to_download_before_save. When this number is reach, stop the loop, save and refresh the TODO
list of URL